+++ lib/sidekiq/cli.rb
@@ -28,7 +28,7 @@ module Sidekiq
     include Singleton
 
     # Used for CLI testing
-    attr_accessor :code, :manager
+    attr_accessor :code
 
     def initialize
       @code = nil
@@ -44,6 +44,7 @@ module Sidekiq
       options.merge!(config.merge(cli))
 
       Sidekiq::Util.logger.level = Logger::DEBUG if options[:verbose]
+      Celluloid.logger = nil
 
       validate!
       write_pid
@@ -51,13 +52,16 @@ module Sidekiq
     end
 
     def run
-      @manager = Sidekiq::Manager.new(options)
+      manager = Sidekiq::Manager.new(options)
+      poller = Sidekiq::Retry::Poller.new
       begin
         logger.info 'Starting processing, hit Ctrl-C to stop'
         manager.start!
+        poller.poll!
         sleep
       rescue Interrupt
         logger.info 'Shutting down'
+        poller.terminate
         manager.stop!(:shutdown => true, :timeout => options[:timeout])
         manager.wait(:shutdown)
       end