+++ myapp/config/routes.rb
@@ -6,4 +6,5 @@ Myapp::Application.routes.draw do
   get "work/email" => "work#email"
   get "work/post" => "work#delayed_post"
   get "work/long" => "work#long"
+  get "work/crash" => "work#crash"
 end